Splitted code in a bunch of modules
