+++ rebbleHandlers/routehandler.go
@@ -5,13 +5,14 @@ import (
 	"net/http"
 
 	"pebble-dev/rebble-auth/db"
+	"pebble-dev/rebble-auth/sso"
 )
 
 // HandlerContext is our struct for storing the data we want to inject in to each handler
 // we can also add things like authorization level, user information, templates, etc.
 type HandlerContext struct {
 	Database *db.Handler
-	SSos     []Sso
+	SSos     []sso.Sso
 }
 
 // routeHandler is a struct that implements http.Handler, allowing us to inject a custom context
